main_camp/ch4_WebService/DRF/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
autopep8==2.3.1
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
certifi @ file:///C:/b/abs_59o0xj7aav/croot/certifi_1734473304008/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
colorama==0.4.6
decorator==5.1.1
Django==4.2.1
django-extensions==3.2.3
django-seed==0.3.1
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
Faker==33.1.0
fontawesomefree==6.6.0
gprof2dot==2024.6.6
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
ipython==8.30.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pure_eval==0.2.3
pycodestyle==2.12.1
Pygments==2.18.0
PyJWT==2.10.1
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dateutil==2.9.0.post0
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
toposort==1.10
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3 @ file:///C:/b/abs_9a_f8h_bn2/croot/urllib3_1727769836930/work
wcwidth==0.2.13
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
